@@ -1,3 +1,4 @@
+require 'cover_me'
 ENV["RAILS_ENV"] = "test"
 require File.expand_path('../../config/environment', __FILE__)
 require 'rails/test_help'
@@ -11,7 +12,8 @@ class ActiveSupport::TestCase
       id_attributes.each do |attr|
         fixture_json[attr] = BSON::ObjectId.from_string(fixture_json[attr])
       end
-      MONGO_DB[collection].save(fixture_json, :safe=>true)
+      # run in safe mode so that we block until insert completes... otherwise the tests can execute before data is available.
+      MONGO_DB[collection].save(fixture_json, safe: true)
     end
   end
 end
