@@ -11,7 +11,7 @@ QueryComposer::Application.configure do
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
-  config.action_view.debug_rjs             = true
+#  config.action_view.debug_rjs             = true
   config.action_controller.perform_caching = false
 
   # Don't care if the mailer can't send
@@ -24,12 +24,12 @@ QueryComposer::Application.configure do
   config.action_dispatch.best_standards_support = :builtin
 
   # add support for Pry debugging
-  silence_warnings do
-    begin
-      require 'pry'
-      IRB = Pry
-    rescue LoadError
-    end
-  end
+#  silence_warnings do
+#    begin
+#      require 'pry'
+#      IRB = Pry
+#    rescue LoadError
+#    end
+#  end
 end
 
